dataflow-types: move all connection type into connections module
This commit is pure code movement.
