+++ misc/python/materialize/cli/run.py
@@ -126,6 +126,7 @@ def main() -> int:
         for test in args.test:
             command += ["--test", test]
         command += args.args
+        command += ["--", "--nocapture"]
         os.environ["POSTGRES_URL"] = args.postgres
     else:
         raise UIError(f"unknown program {args.program}")
