+++ src/compute/src/bin/computed.rs
@@ -19,8 +19,8 @@ use tracing::info;
 
 use mz_build_info::{build_info, BuildInfo};
 use mz_dataflow_types::client::{ComputeClient, ComputeCommand, ComputeResponse, GenericClient};
+use mz_dataflow_types::connections::ConnectionContext;
 use mz_dataflow_types::reconciliation::command::ComputeCommandReconcile;
-use mz_dataflow_types::ConnectionContext;
 use mz_orchestrator_tracing::TracingCliArgs;
 use mz_ore::cli::{self, CliConfig};
 use mz_ore::metrics::MetricsRegistry;
