+++ src/dataflow-types/src/types/connections/aws.rs
@@ -18,7 +18,10 @@ use mz_repr::proto::{IntoRustIfSome, ProtoType, RustType, TryFromProtoError};
 use mz_repr::url::URL_PATTERN;
 use mz_repr::GlobalId;
 
-include!(concat!(env!("OUT_DIR"), "/mz_dataflow_types.types.aws.rs"));
+include!(concat!(
+    env!("OUT_DIR"),
+    "/mz_dataflow_types.types.connections.aws.rs"
+));
 
 /// A wrapper for [`Uri`] that implements [`Serialize`] and `Deserialize`.
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -64,7 +67,7 @@ impl RustType<ProtoSerdeUri> for SerdeUri {
 /// `AwsExternalIdPrefix`. The only approved way to construct an `AwsExternalIdPrefix`
 /// is via [`ConnectionContext::from_cli_args`].
 ///
-/// [`ConnectionContext::from_cli_args`]: crate::ConnectionContext::from_cli_args
+/// [`ConnectionContext::from_cli_args`]: crate::connections::ConnectionContext::from_cli_args
 /// [external ID]: https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_create_for-user_externalid.html
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
 pub struct AwsExternalIdPrefix(pub(super) String);
