+++ src/sql/src/plan/statement/show.rs
@@ -393,7 +393,7 @@ fn show_sources<'a>(
                  mz_internal.mz_classify_object_id(id) AS type,
                  mz_internal.mz_is_materialized(id) AS materialized,
                  volatility,
-                 connection_type
+                 type
              FROM
                  mz_catalog.mz_sources
              WHERE
@@ -405,7 +405,7 @@ fn show_sources<'a>(
                  name,
                  mz_internal.mz_classify_object_id(id) AS type,
                  volatility,
-                 connection_type
+                 type
              FROM
                  mz_catalog.mz_sources
              WHERE
